requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.12.1
backports.functools-lru-cache==1.5
billiard==3.3.0.23
celery==3.1.26.post2
Click==7.0
dateparser==0.7.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.17
django-admin-interface==0.8.2
django-advanced-filters==1.1.0
django-appconf==1.0.2
django-avatar==4.1.0
django-bootstrap3==11.0.0
django-braces==1.13.0
django-calendarium==1.3.4
django-celery==3.2.2
django-colorfield==0.1.15
django-crispy-forms==1.7.2
django-datatable==0.3.1
django-easyfilters==0.6
django-extensions==2.1.4
django-filer==1.3.2
django-filter==1.1.0
django-filters==0.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==1.1.0
django-js-asset==1.1.0
django-libs==2.0.2
django-mptt==0.9.1
django-polymorphic==2.0.3
django-tables==0.2
django-tables2==2.0.2
django-widget-tweaks==1.4.3
django-wpadmin==1.7.4
djangorestframework==3.9.0
djangorestframework-filters==0.10.2.post0
easy-thumbnails==2.5
et-xmlfile==1.0.1
future==0.17.1
jdcal==1.4
kombu==3.0.37
odfpy==1.3.6
openpyxl==2.5.9
Pillow==5.3.0
psycopg2==2.6.2
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
regex==2018.11.22
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
typing==3.6.6
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
xlrd==1.1.0
xlwt==1.3.0
